requirements.txt
langchain
python-dotenv
langchain_community
langchain-text-splitters
langchain-openai
chromadb
feature-engine
openai
sentence_transformers
faiss-cpu
langchain_huggingface
langchain_chroma
streamlit
langchain_groq
bs4
duckdb
pymupdf
arxiv
wikipedia
SQLAlchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
cassio
datasets
langchain
openai
tiktoken
gradio
langchain_nvidia_ai_endpoints
faiss-cpu
crewai
crewai_tools
pinecone
pinecone-text
pinecone-notebooks
neo4j==5.14